// src/canisters/yuku-old/yuku_core/module_8d3a5afa/core_8d3a5afa.did.d.ts
export type CandyValue =
    | { Int: bigint }
    | { Nat: bigint }
    | { Empty: null }
    | { Nat16: number }
    | { Nat32: number }
    | { Nat64: bigint }
    | { Blob: Uint8Array | number[] }
    | { Bool: boolean }
    | { Int8: number }
    | { Nat8: number }
    | { Nats: { thawed: Array<bigint> } | { frozen: Array<bigint> } }
    | { Text: string }
    | {
          Bytes: { thawed: Uint8Array | number[] } | { frozen: Uint8Array | number[] };
      }
    | { Int16: number }
    | { Int32: number }
    | { Int64: bigint }
    | { Option: [] | [CandyValue] }
    | { Floats: { thawed: Array<number> } | { frozen: Array<number> } }
    | { Float: number }
    | { Principal: Principal }
    | {
          Array: { thawed: Array<CandyValue> } | { frozen: Array<CandyValue> };
      }
    | { Class: Array<Property> };
// ...
export interface ICTokenSpec {
    fee: bigint;
    decimals: bigint;
    canister: Principal;
    standard: { ICRC1: null } | { EXTFungible: null } | { DIP20: null } | { Ledger: null };
    symbol: string;
}
// ...
export interface OgyInfo {
    fee: { rate: bigint; precision: bigint };
    creator: Principal;
    token: TokenSpec__1;
    owner: Principal;
    totalFee: { rate: bigint; precision: bigint };
}
// ...
export type Standard = { ext: null } | { ogy: OgyInfo };
// ...
export type TokenSpec__1 = { ic: ICTokenSpec } | { extensible: CandyValue };